Separate sanitization and validation to request-builder.js
